+++ src/utils/message-channel-utils.ts
@@ -1,7 +1,9 @@
 import { tinyassert } from "@hiogawa/utils";
 import { generateId } from "./misc";
 
+//
 // instantiate MessageChannel in preload and send MessagePort to main
+//
 
 export async function shareMessagePortPreload(
   ipcRenderer: Electron.IpcRenderer,
@@ -55,4 +57,76 @@ export function receiveMessagePortMainPromise(
       resolve(port);
     });
   });
+}
+
+//
+// fix up small interface differences of MessagePort to be uniformally usable in main/preload/renderer
+//
+
+export function normalizeMessagePortMain(
+  port: Electron.MessagePortMain
+): MessagePort {
+  const listerWrappers = new WeakMap<object, any>();
+
+  return {
+    start: port.start.bind(port),
+    close: port.close.bind(port),
+
+    // @ts-expect-error inessential type imcompatibility
+    postMessage: (message: any, transfer: Transferable[] = []) => {
+      tinyassert(transfer.length === 0);
+      port.postMessage(message, []);
+    },
+
+    addEventListener: (type: string, listener: any) => {
+      tinyassert(type === "message");
+      const wrapper = (event: Electron.MessageEvent) => {
+        listener({ data: event.data } as MessageEvent);
+      };
+      port.on("message", wrapper);
+      listerWrappers.set(listener, wrapper);
+    },
+
+    removeEventListener: (type: string, listener: any) => {
+      tinyassert(type === "message");
+      const wrapper = listerWrappers.get(listener);
+      if (wrapper) {
+        port.off("message", wrapper);
+        listerWrappers.delete(listener);
+      }
+    },
+  };
+}
+
+export function normalizeMessagePortPreload(port: MessagePort): MessagePort {
+  const listerWrappers = new WeakMap<object, any>();
+
+  return {
+    start: port.start.bind(port),
+    close: port.close.bind(port),
+
+    // @ts-expect-error inessential type imcompatibility
+    postMessage: (message: any, transfer: Transferable[] = []) => {
+      tinyassert(transfer.length === 0);
+      port.postMessage(message, []);
+    },
+
+    addEventListener: (type: string, listener: any) => {
+      tinyassert(type === "message");
+      const wrapper = (event: MessageEvent) => {
+        listener({ data: event.data }); // strip out non "data" properties since there seem to be something non-serializable/proxy-able.
+      };
+      port.addEventListener(type, wrapper);
+      listerWrappers.set(listener, wrapper);
+    },
+
+    removeEventListener: (type: string, listener: any) => {
+      tinyassert(type === "message");
+      const wrapper = listerWrappers.get(listener);
+      if (wrapper) {
+        port.removeEventListener(type, wrapper);
+        listerWrappers.delete(listener);
+      }
+    },
+  };
 }
